test: move custom errors into seperate test file `errors.ts`
