+++ tests/LSP6KeyManager/internals/AllowedERC725YKeys.internal.ts
@@ -14,7 +14,10 @@ import { LSP6InternalsTestContext } from "../../utils/context";
 import { setupKeyManagerHelper } from "../../utils/fixtures";
 
 // helpers
-import { abiCoder, NotAllowedERC725YKeyError } from "../../utils/helpers";
+import { abiCoder } from "../../utils/helpers";
+
+// errors
+import { NotAllowedERC725YKeyError } from "../../utils/errors";
 
 export const testAllowedERC725YKeysInternals = (
   buildContext: () => Promise<LSP6InternalsTestContext>
