+++ tests/LSP6KeyManager/tests/PermissionDeploy.test.ts
@@ -17,8 +17,8 @@ import {
 import { LSP6TestContext } from "../../utils/context";
 import { setupKeyManager } from "../../utils/fixtures";
 
-// helpers
-import { NotAuthorisedError } from "../../utils/helpers";
+// errors
+import { NotAuthorisedError } from "../../utils/errors";
 
 export const shouldBehaveLikePermissionDeploy = (
   buildContext: () => Promise<LSP6TestContext>
