+++ tests/LSP6KeyManager/tests/PermissionSetData.test.ts
@@ -22,8 +22,10 @@ import {
   abiCoder,
   generateKeysAndValues,
   getRandomAddresses,
-  NotAuthorisedError,
 } from "../../utils/helpers";
+
+// errors
+import { NotAuthorisedError } from "../../utils/errors";
 
 export const shouldBehaveLikePermissionSetData = (
   buildContext: () => Promise<LSP6TestContext>
